Improve custom responses with Boom
