--- packages/strapi/lib/configuration/hooks/core/responses/responses/forbidden.js
@@ -1,19 +1,0 @@
-'use strict';
-
-/**
- * Default `forbidden` response.
- */
-
-module.exports = function forbidden(data) {
-
-  // Set the status.
-  this.status = 403;
-
-  // Delete the `data` object if the app is used in production environment.
-  if (strapi.config.environment === 'production') {
-    data = '';
-  }
-
-  // Finally send the response.
-  this.body = data;
-};
