+++ packages/strapi/lib/configuration/hooks/core/responses/responses/index.js
@@ -5,10 +5,6 @@
  */
 
 module.exports = {
-  badRequest: require('./badRequest'),
   created: require('./created'),
-  forbidden: require('./forbidden'),
-  notFound: require('./notFound'),
-  ok: require('./ok'),
-  serverError: require('./serverError')
+  send: require('./send')
 };
