--- packages/strapi/lib/configuration/hooks/core/responses/responses/notFound.js
@@ -1,19 +1,0 @@
-'use strict';
-
-/**
- * Default `notFound` response.
- */
-
-module.exports = function notFound(data) {
-
-  // Set the status.
-  this.status = 404;
-
-  // Delete the `data` object if the app is used in production environment.
-  if (strapi.config.environment === 'production') {
-    data = '';
-  }
-
-  // Finally send the response.
-  this.body = data;
-};
